requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
openai==1.75.0
requests==2.32.3
python-dotenv==1.1.0
pydantic==2.11.3
